+++ neoforge/src/main/kotlin/com/cobblemonrei/neoforge/CobblemonSpawningNeoForgeEMI.kt
@@ -1,3 +1,4 @@
 package com.cobblemonrei.neoforge
 
+@dev.emi.emi.api.EmiEntrypoint
 class CobblemonSpawningNeoForgeEMI : com.cobblemonrei.emi.CobblemonEMIPlugin()